smashedavo/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.13
django-bootstrap-v5==1.0.11
django-environ==0.11.2
django-formset==1.4.4
psycopg==3.1.19
psycopg-binary==3.1.19
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
